[Refactor]	 Transformation moved into transformation.base
