+++ src/finn/transformation/fpgadataflow/cleanup.py
@@ -31,7 +31,7 @@ import shutil
 
 import finn.custom_op.registry as registry
 from finn.util.fpgadataflow import is_fpgadataflow_node
-from finn.transformation import Transformation
+from finn.transformation.base import Transformation
 
 
 class CleanUp(Transformation):
