+++ src/finn/transformation/fpgadataflow/hlssynth_ip.py
@@ -29,7 +29,7 @@
 import os
 import finn.custom_op.registry as registry
 from finn.util.fpgadataflow import is_fpgadataflow_node
-from finn.transformation import NodeLocalTransformation
+from finn.transformation.base import NodeLocalTransformation
 import warnings
