+++ src/finn/transformation/fpgadataflow/insert_tlastmarker.py
@@ -30,7 +30,7 @@ from onnx import TensorProto
 from onnx import helper as oh
 
 from finn.custom_op.registry import getCustomOp
-from finn.transformation import Transformation
+from finn.transformation.base import Transformation
 from finn.util.basic import get_by_name
 
 import numpy as np
