+++ src/finn/transformation/fpgadataflow/make_deployment.py
@@ -32,7 +32,7 @@ import subprocess
 from distutils.dir_util import copy_tree
 from shutil import copy
 
-from finn.transformation import Transformation
+from finn.transformation.base import Transformation
 from finn.util.basic import make_build_dir
 import finn.transformation.fpgadataflow.templates as templates
