+++ src/finn/transformation/fpgadataflow/prepare_cppsim.py
@@ -31,7 +31,7 @@ import os
 import finn.custom_op.registry as registry
 from finn.util.basic import make_build_dir
 from finn.util.fpgadataflow import is_fpgadataflow_node
-from finn.transformation import Transformation
+from finn.transformation.base import Transformation
 from finn.util.basic import get_num_default_workers
 import multiprocessing as mp
 import copy
