+++ src/finn/transformation/fpgadataflow/prepare_ip.py
@@ -28,7 +28,7 @@
 
 import os
 import finn.custom_op.registry as registry
-from finn.transformation import Transformation
+from finn.transformation.base import Transformation
 from finn.util.basic import make_build_dir
 from finn.util.fpgadataflow import is_fpgadataflow_node
 import warnings
