+++ src/finn/transformation/fpgadataflow/replace_verilog_relpaths.py
@@ -30,7 +30,7 @@ import os
 
 import finn.custom_op.registry as registry
 from finn.util.fpgadataflow import is_fpgadataflow_node
-from finn.transformation import Transformation
+from finn.transformation.base import Transformation
 
 
 class ReplaceVerilogRelPaths(Transformation):
