+++ src/finn/transformation/fpgadataflow/set_exec_mode.py
@@ -28,7 +28,7 @@
 
 import finn.custom_op.registry as registry
 from finn.util.fpgadataflow import is_fpgadataflow_node
-from finn.transformation import Transformation
+from finn.transformation.base import Transformation
 
 
 class SetExecMode(Transformation):
