+++ src/finn/transformation/fpgadataflow/synth_ooc.py
@@ -29,7 +29,7 @@
 import os
 from shutil import copy2
 
-from finn.transformation import Transformation
+from finn.transformation.base import Transformation
 from finn.util.vivado import out_of_context_synth
 from finn.util.basic import make_build_dir
