+++ src/finn/transformation/streamline/collapse_repeated.py
@@ -28,7 +28,7 @@
 
 from onnx import helper as oh
 
-from finn.transformation import Transformation
+from finn.transformation.base import Transformation
 from finn.transformation.infer_shapes import InferShapes
 from finn.core.datatype import DataType
