+++ src/finn/transformation/streamline/sign_to_thres.py
@@ -30,7 +30,7 @@ import numpy as np
 from onnx import helper as oh
 
 from finn.core.datatype import DataType
-from finn.transformation import Transformation
+from finn.transformation.base import Transformation
 
 
 class ConvertSignToThres(Transformation):
